[BACKLOG-39945] Dashboards - Unable to get the dropdown of prompts to assigning Sources to Parameters
The previous version of this code needlessly had a local variable to store whether the popup was showing or not. The only advantage was that it allowed to test for the popup to be opened without having to know if the popup had already been created. However, it was redundantly implementing logic already present in the PopupPanel class.

The code failed to keep the local state in sync with the popup's actual state, in the case where the dialog was closed during the process of showing it (via `togglePopup()`), due to being detected the anchor component was off view.